requirements.txt
fastapi
uvicorn
onnxruntime
numpy
pydantic
librosa
soundfile
Pillow
python-multipart